# decode: Reduce allocations, memory usage & table size